briefing-enhancer-service/requirements.txt
fastapi==0.115.0
fastapi[security]==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.0
sqlalchemy>=2.0.36
alembic==1.13.1
psycopg2-binary>=2.9.10
langchain>=0.1.13
langchain-openai>=0.0.6
langchain-google-genai>=0.0.6
langchain-huggingface>=0.0.1
langgraph>=0.0.40
langgraph-checkpoint-postgres>=2.0.0
openai>=1.12.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.25.0
psycopg[binary]>=3.1.0
pyyaml>=6.0.1
redis>=5.0.0
prometheus-client>=0.21.0
prometheus-fastapi-instrumentator>=7.0.0

# Tracing
langsmith>=0.1.0
